like button refactor
